Renamed float_wrapper.py to json_util.py, and moved it to the new directory tensorboard/lib/python directory for shared Python utilities. Change: 118634201
