+++ tensorflow/tensorboard/lib/python/BUILD
@@ -1,0 +1,36 @@
+# Description:
+# Public utilities from TensorBoard, a dashboard for investigating TensorFlow
+
+package(default_visibility = ["//tensorflow:internal"])
+
+licenses(["notice"])  # Apache 2.0
+
+exports_files(["LICENSE"])
+
+py_library(
+    name = "json_util",
+    srcs = ["json_util.py"],
+    srcs_version = "PY2AND3",
+    visibility = ["//visibility:public"],
+)
+
+py_test(
+    name = "json_util_test",
+    size = "small",
+    srcs = ["json_util_test.py"],
+    srcs_version = "PY2AND3",
+    deps = [
+        ":json_util",
+        "//tensorflow/python:platform_test",
+    ],
+)
+
+filegroup(
+    name = "all_files",
+    srcs = glob(
+        [
+            "*",
+        ],
+    ),
+    visibility = ["//tensorflow:__subpackages__"],
+)
